pyproject.toml
[project]
name = "booksyncbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.24.0",
    "asyncpg>=0.30.0",
    "google-api-python-client>=2.187.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.2.3",
]

[dependency-groups]
dev = [
    "ruff>=0.14.10",
    "ty>=0.0.9",
    "watchfiles>=1.0.0",
]
